Split up.go in different files
