+++ cmd/up/up.go
@@ -14,7 +14,6 @@
 package up
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"os"
@@ -32,26 +31,17 @@ import (
 	"github.com/okteto/okteto/pkg/errors"
 	k8Client "github.com/okteto/okteto/pkg/k8s/client"
 	"github.com/okteto/okteto/pkg/k8s/deployments"
-	"github.com/okteto/okteto/pkg/k8s/exec"
-	okLabels "github.com/okteto/okteto/pkg/k8s/labels"
 	"github.com/okteto/okteto/pkg/k8s/namespaces"
-	"github.com/okteto/okteto/pkg/k8s/pods"
-	"github.com/okteto/okteto/pkg/k8s/secrets"
-	"github.com/okteto/okteto/pkg/k8s/services"
-	"github.com/okteto/okteto/pkg/k8s/volumes"
 	"github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
 	"github.com/okteto/okteto/pkg/okteto"
 	"github.com/okteto/okteto/pkg/registry"
 	"github.com/okteto/okteto/pkg/ssh"
 
-	"github.com/okteto/okteto/pkg/k8s/forward"
 	"github.com/okteto/okteto/pkg/syncthing"
 
 	"github.com/spf13/cobra"
 	appsv1 "k8s.io/api/apps/v1"
-	apiv1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // ReconnectingMessage is the message shown when we are trying to reconnect
@@ -311,159 +301,6 @@ func (up *upContext) activateLoop(autoDeploy, build bool) {
 	}
 }
 
-func (up *upContext) activate(autoDeploy, build bool) error {
-	log.Infof("activating development container retry=%t", up.isRetry)
-
-	if err := config.UpdateStateFile(up.Dev, config.Activating); err != nil {
-		return err
-	}
-
-	// create a new context on every iteration
-	ctx, cancel := context.WithCancel(context.Background())
-	up.Cancel = cancel
-	up.ShutdownCompleted = make(chan bool, 1)
-	up.Sy = nil
-	up.Forwarder = nil
-	defer up.shutdown()
-
-	up.Disconnect = make(chan error, 1)
-	up.CommandResult = make(chan error, 1)
-	up.cleaned = make(chan string, 1)
-	up.hardTerminate = make(chan error, 1)
-
-	d, create, err := up.getCurrentDeployment(ctx, autoDeploy)
-	if err != nil {
-		return err
-	}
-
-	if up.isRetry && !deployments.IsDevModeOn(d) {
-		log.Information("Development container has been deactivated")
-		return nil
-	}
-
-	if deployments.IsDevModeOn(d) && deployments.HasBeenChanged(d) {
-		return errors.UserError{
-			E: fmt.Errorf("Deployment '%s' has been modified while your development container was active", d.Name),
-			Hint: `Follow these steps:
-	  1. Execute 'okteto down'
-	  2. Apply your manifest changes again: 'kubectl apply'
-	  3. Execute 'okteto up' again
-    More information is available here: https://okteto.com/docs/reference/known-issues/index.html#kubectl-apply-changes-are-undone-by-okteto-up`,
-		}
-	}
-
-	if _, err := registry.GetImageTagWithDigest(ctx, up.Dev.Namespace, up.Dev.Image.Name); err == errors.ErrNotFound {
-		log.Infof("image '%s' not found, building it: %s", up.Dev.Image.Name, err.Error())
-		build = true
-	}
-
-	if !up.isRetry && build {
-		if err := up.buildDevImage(ctx, d, create); err != nil {
-			return fmt.Errorf("error building dev image: %s", err)
-		}
-	}
-
-	go up.initializeSyncthing()
-
-	if err := up.setDevContainer(d); err != nil {
-		return err
-	}
-
-	if err := up.devMode(ctx, d, create); err != nil {
-		if errors.IsTransient(err) {
-			return err
-		}
-		return fmt.Errorf("couldn't activate your development container\n    %s", err.Error())
-	}
-
-	log.Success("Development container activated")
-	up.isRetry = true
-
-	if err := up.forwards(ctx); err != nil {
-		if err == errors.ErrSSHConnectError {
-			err := up.checkOktetoStartError(ctx, "Failed to connect to your development container")
-			if err == errors.ErrLostSyncthing {
-				if err := pods.Destroy(ctx, up.Pod.Name, up.Dev.Namespace, up.Client); err != nil {
-					return fmt.Errorf("error recreating development container: %s", err.Error())
-				}
-			}
-			return err
-		}
-		return fmt.Errorf("couldn't connect to your development container: %s", err.Error())
-	}
-	log.Success("Connected to your development container")
-
-	go up.cleanCommand(ctx)
-
-	if err := up.sync(ctx); err != nil {
-		if up.shouldRetry(ctx, err) {
-			return errors.ErrLostSyncthing
-		}
-		return err
-	}
-
-	up.success = true
-	if up.isRetry {
-		analytics.TrackReconnect(true, up.isSwap)
-	}
-	log.Success("Files synchronized")
-
-	go func() {
-		output := <-up.cleaned
-		log.Debugf("clean command output: %s", output)
-
-		outByCommand := strings.Split(strings.TrimSpace(output), "\n")
-		if len(outByCommand) >= 2 {
-			version, watches := outByCommand[0], outByCommand[1]
-
-			if isWatchesConfigurationTooLow(watches) {
-				folder := config.GetNamespaceHome(up.Dev.Namespace)
-				if utils.GetWarningState(folder, ".remotewatcher") == "" {
-					log.Yellow("The value of /proc/sys/fs/inotify/max_user_watches in your cluster nodes is too low.")
-					log.Yellow("This can affect file synchronization performance.")
-					log.Yellow("Visit https://okteto.com/docs/reference/known-issues/index.html for more information.")
-					if err := utils.SetWarningState(folder, ".remotewatcher", "true"); err != nil {
-						log.Infof("failed to set warning remotewatcher state: %s", err.Error())
-					}
-				}
-			}
-
-			if version != model.OktetoBinImageTag {
-				log.Yellow("The Okteto CLI version %s uses the init container image %s.", config.VersionString, model.OktetoBinImageTag)
-				log.Yellow("Please consider upgrading your init container image %s with the content of %s", up.Dev.InitContainer.Image, model.OktetoBinImageTag)
-				log.Infof("Using init image %s instead of default init image (%s)", up.Dev.InitContainer.Image, model.OktetoBinImageTag)
-			}
-		}
-		printDisplayContext(up.Dev)
-		up.CommandResult <- up.runCommand(ctx)
-	}()
-	prevError := up.waitUntilExitOrInterrupt()
-
-	if up.shouldRetry(ctx, prevError) {
-		if !up.Dev.PersistentVolumeEnabled() {
-			if err := pods.Destroy(ctx, up.Pod.Name, up.Dev.Namespace, up.Client); err != nil {
-				return err
-			}
-		}
-		return errors.ErrLostSyncthing
-	}
-
-	return prevError
-}
-
-func (up *upContext) shouldRetry(ctx context.Context, err error) bool {
-	switch err {
-	case nil:
-		return false
-	case errors.ErrLostSyncthing:
-		return true
-	case errors.ErrCommandFailed:
-		return !up.Sy.Ping(ctx, false)
-	}
-
-	return false
-}
-
 func (up *upContext) getCurrentDeployment(ctx context.Context, autoDeploy bool) (*appsv1.Deployment, bool, error) {
 	d, err := deployments.Get(ctx, up.Dev, up.Dev.Namespace, up.Client)
 	if err == nil {
@@ -587,481 +424,6 @@ func (up *upContext) setDevContainer(d *appsv1.Deployment) error {
 	}
 
 	return nil
-}
-
-func (up *upContext) devMode(ctx context.Context, d *appsv1.Deployment, create bool) error {
-	if err := up.createDevContainer(ctx, d, create); err != nil {
-		return err
-	}
-
-	return up.waitUntilDevelopmentContainerIsRunning(ctx)
-}
-
-func (up *upContext) createDevContainer(ctx context.Context, d *appsv1.Deployment, create bool) error {
-	spinner := utils.NewSpinner("Activating your development container...")
-	spinner.Start()
-	defer spinner.Stop()
-
-	if err := config.UpdateStateFile(up.Dev, config.Starting); err != nil {
-		return err
-	}
-
-	if up.Dev.PersistentVolumeEnabled() {
-		if err := volumes.Create(ctx, up.Dev, up.Client); err != nil {
-			return err
-		}
-	}
-
-	trList, err := deployments.GetTranslations(ctx, up.Dev, d, up.Client)
-	if err != nil {
-		return err
-	}
-
-	if err := deployments.TranslateDevMode(trList, up.Client, up.isOktetoNamespace); err != nil {
-		return err
-	}
-
-	initSyncErr := <-up.hardTerminate
-	if initSyncErr != nil {
-		return initSyncErr
-	}
-
-	log.Info("create deployment secrets")
-	if err := secrets.Create(ctx, up.Dev, up.Client, up.Sy); err != nil {
-		return err
-	}
-
-	for name := range trList {
-		if name == d.Name {
-			if err := deployments.Deploy(ctx, trList[name].Deployment, create, up.Client); err != nil {
-				return err
-			}
-		} else {
-			if err := deployments.Deploy(ctx, trList[name].Deployment, false, up.Client); err != nil {
-				return err
-			}
-		}
-
-		if trList[name].Deployment.Annotations[okLabels.DeploymentAnnotation] == "" {
-			continue
-		}
-
-		if err := deployments.UpdateOktetoRevision(ctx, trList[name].Deployment, up.Client); err != nil {
-			return err
-		}
-
-	}
-
-	if create {
-		if err := services.CreateDev(ctx, up.Dev, up.Client); err != nil {
-			return err
-		}
-	}
-
-	pod, err := pods.GetDevPodInLoop(ctx, up.Dev, up.Client, create)
-	if err != nil {
-		return err
-	}
-
-	up.Pod = pod
-	return nil
-}
-
-func (up *upContext) waitUntilDevelopmentContainerIsRunning(ctx context.Context) error {
-	spinner := utils.NewSpinner("Activating your development container...")
-	spinner.Start()
-	defer spinner.Stop()
-
-	optsWatchPod := metav1.ListOptions{
-		Watch:         true,
-		FieldSelector: fmt.Sprintf("metadata.name=%s", up.Pod.Name),
-	}
-
-	watcherPod, err := up.Client.CoreV1().Pods(up.Dev.Namespace).Watch(ctx, optsWatchPod)
-	if err != nil {
-		return err
-	}
-
-	optsWatchEvents := metav1.ListOptions{
-		Watch:         true,
-		FieldSelector: fmt.Sprintf("involvedObject.kind=Pod,involvedObject.name=%s", up.Pod.Name),
-	}
-
-	watcherEvents, err := up.Client.CoreV1().Events(up.Dev.Namespace).Watch(ctx, optsWatchEvents)
-	if err != nil {
-		return err
-	}
-
-	if up.Dev.PersistentVolumeEnabled() {
-		spinner.Update("Activating: attaching persistent volume...")
-		if err := config.UpdateStateFile(up.Dev, config.Attaching); err != nil {
-			log.Infof("error updating state: %s", err.Error())
-		}
-	}
-
-	for {
-		select {
-		case event := <-watcherEvents.ResultChan():
-			e, ok := event.Object.(*apiv1.Event)
-			if !ok {
-				watcherEvents, err = up.Client.CoreV1().Events(up.Dev.Namespace).Watch(ctx, optsWatchEvents)
-				if err != nil {
-					return err
-				}
-				continue
-			}
-			log.Infof("pod %s event: %s", up.Pod, e.Message)
-			optsWatchEvents.ResourceVersion = e.ResourceVersion
-			switch e.Reason {
-			case "Failed", "FailedScheduling", "FailedCreatePodSandBox", "ErrImageNeverPull", "InspectFailed", "FailedCreatePodContainer":
-				if strings.Contains(e.Message, "pod has unbound immediate PersistentVolumeClaims") {
-					continue
-				}
-				return fmt.Errorf(e.Message)
-			case "SuccessfulAttachVolume":
-				spinner.Update("Activating: pulling images...")
-			case "Killing":
-				return errors.ErrDevPodDeleted
-			case "Pulling":
-				message := strings.Replace(e.Message, "Pulling", "pulling", 1)
-				spinner.Update(fmt.Sprintf("Activating: %s...", message))
-				if err := config.UpdateStateFile(up.Dev, config.Pulling); err != nil {
-					log.Infof("error updating state: %s", err.Error())
-				}
-			}
-		case event := <-watcherPod.ResultChan():
-			pod, ok := event.Object.(*apiv1.Pod)
-			if !ok {
-				watcherPod, err = up.Client.CoreV1().Pods(up.Dev.Namespace).Watch(ctx, optsWatchPod)
-				if err != nil {
-					return err
-				}
-				continue
-			}
-			log.Infof("dev pod %s is now %s", pod.Name, pod.Status.Phase)
-			if pod.Status.Phase == apiv1.PodRunning {
-				return nil
-			}
-			if pod.DeletionTimestamp != nil {
-				return errors.ErrDevPodDeleted
-			}
-		case <-ctx.Done():
-			log.Debug("call to waitUntilDevelopmentContainerIsRunning cancelled")
-			return ctx.Err()
-		}
-	}
-}
-
-func (up *upContext) forwards(ctx context.Context) error {
-	spinner := utils.NewSpinner("Connecting to your development container...")
-	spinner.Start()
-	defer spinner.Stop()
-
-	if up.Dev.RemoteModeEnabled() {
-		return up.sshForwards(ctx)
-	}
-
-	log.Infof("starting port forwards")
-	up.Forwarder = forward.NewPortForwardManager(ctx, up.Dev.Interface, up.RestConfig, up.Client, up.Dev.Namespace)
-
-	for idx, f := range up.Dev.Forward {
-		if f.Labels != nil {
-			forwardWithServiceName, err := up.Forwarder.TransformLabelsToServiceName(f)
-			if err != nil {
-				return err
-			}
-			up.Dev.Forward[idx] = forwardWithServiceName
-			f = forwardWithServiceName
-		}
-		if err := up.Forwarder.Add(f); err != nil {
-			return err
-		}
-	}
-
-	if err := up.Forwarder.Add(model.Forward{Local: up.Sy.RemotePort, Remote: syncthing.ClusterPort}); err != nil {
-		return err
-	}
-
-	if err := up.Forwarder.Add(model.Forward{Local: up.Sy.RemoteGUIPort, Remote: syncthing.GUIPort}); err != nil {
-		return err
-	}
-
-	return up.Forwarder.Start(up.Pod.Name, up.Dev.Namespace)
-}
-
-func (up *upContext) sshForwards(ctx context.Context) error {
-	log.Infof("starting SSH port forwards")
-	f := forward.NewPortForwardManager(ctx, up.Dev.Interface, up.RestConfig, up.Client, up.Dev.Namespace)
-	if err := f.Add(model.Forward{Local: up.Dev.RemotePort, Remote: up.Dev.SSHServerPort}); err != nil {
-		return err
-	}
-
-	up.Forwarder = ssh.NewForwardManager(ctx, fmt.Sprintf(":%d", up.Dev.RemotePort), up.Dev.Interface, "0.0.0.0", f, up.Dev.Namespace)
-
-	if err := up.Forwarder.Add(model.Forward{Local: up.Sy.RemotePort, Remote: syncthing.ClusterPort}); err != nil {
-		return err
-	}
-
-	if err := up.Forwarder.Add(model.Forward{Local: up.Sy.RemoteGUIPort, Remote: syncthing.GUIPort}); err != nil {
-		return err
-	}
-
-	for idx, f := range up.Dev.Forward {
-		if f.Labels != nil {
-			forwardWithServiceName, err := up.Forwarder.TransformLabelsToServiceName(f)
-			if err != nil {
-				return err
-			}
-			up.Dev.Forward[idx] = forwardWithServiceName
-			f = forwardWithServiceName
-		}
-		if err := up.Forwarder.Add(f); err != nil {
-			return err
-		}
-	}
-
-	for _, r := range up.Dev.Reverse {
-		if err := up.Forwarder.AddReverse(r); err != nil {
-			return err
-		}
-	}
-
-	if err := ssh.AddEntry(up.Dev.Name, up.Dev.Interface, up.Dev.RemotePort); err != nil {
-		log.Infof("failed to add entry to your SSH config file: %s", err)
-		return fmt.Errorf("failed to add entry to your SSH config file")
-	}
-
-	return up.Forwarder.Start(up.Pod.Name, up.Dev.Namespace)
-}
-
-func (up *upContext) initializeSyncthing() error {
-	sy, err := syncthing.New(up.Dev)
-	if err != nil {
-		return err
-	}
-
-	up.Sy = sy
-
-	log.Infof("local syncthing initialized: gui -> %d, sync -> %d", up.Sy.LocalGUIPort, up.Sy.LocalPort)
-	log.Infof("remote syncthing initialized: gui -> %d, sync -> %d", up.Sy.RemoteGUIPort, up.Sy.RemotePort)
-
-	if err := up.Sy.SaveConfig(up.Dev); err != nil {
-		log.Infof("error saving syncthing object: %s", err)
-	}
-
-	up.hardTerminate <- up.Sy.HardTerminate()
-
-	return nil
-}
-
-func (up *upContext) sync(ctx context.Context) error {
-	if err := up.startSyncthing(ctx); err != nil {
-		return err
-	}
-
-	return up.synchronizeFiles(ctx)
-}
-
-func (up *upContext) startSyncthing(ctx context.Context) error {
-	spinner := utils.NewSpinner("Starting the file synchronization service...")
-	spinner.Start()
-	if err := config.UpdateStateFile(up.Dev, config.StartingSync); err != nil {
-		return err
-	}
-	defer spinner.Stop()
-
-	if err := up.Sy.Run(ctx); err != nil {
-		return err
-	}
-
-	if err := up.Sy.WaitForPing(ctx, true); err != nil {
-		return err
-	}
-
-	if err := up.Sy.WaitForPing(ctx, false); err != nil {
-		log.Infof("failed to ping syncthing: %s", err.Error())
-		err = up.checkOktetoStartError(ctx, "Failed to connect to the synchronization service")
-		if err == errors.ErrLostSyncthing {
-			if err := pods.Destroy(ctx, up.Pod.Name, up.Dev.Namespace, up.Client); err != nil {
-				return fmt.Errorf("error recreating development container: %s", err.Error())
-			}
-		}
-		return err
-	}
-
-	if up.resetSyncthing {
-		spinner.Update("Resetting synchronization service database...")
-		if err := up.Sy.ResetDatabase(ctx, up.Dev, false); err != nil {
-			return err
-		}
-		if err := up.Sy.ResetDatabase(ctx, up.Dev, true); err != nil {
-			return err
-		}
-
-		if err := up.Sy.WaitForPing(ctx, false); err != nil {
-			return err
-		}
-		if err := up.Sy.WaitForPing(ctx, true); err != nil {
-			return err
-		}
-
-		up.resetSyncthing = false
-	}
-
-	up.Sy.SendStignoreFile(ctx)
-	spinner.Update("Scanning file system...")
-	if err := up.Sy.WaitForScanning(ctx, up.Dev, true); err != nil {
-		return err
-	}
-
-	if !up.Dev.PersistentVolumeEnabled() {
-		if err := up.Sy.WaitForScanning(ctx, up.Dev, false); err != nil {
-			log.Infof("failed to wait for syncthing scanning: %s", err.Error())
-			return up.checkOktetoStartError(ctx, "Failed to connect to the synchronization service")
-		}
-	}
-
-	return nil
-}
-
-func (up *upContext) synchronizeFiles(ctx context.Context) error {
-	suffix := "Synchronizing your files..."
-	spinner := utils.NewSpinner(suffix)
-	pbScaling := 0.30
-
-	if err := config.UpdateStateFile(up.Dev, config.Synchronizing); err != nil {
-		return err
-	}
-	spinner.Start()
-	defer spinner.Stop()
-	reporter := make(chan float64)
-	go func() {
-		<-time.NewTicker(2 * time.Second).C
-		var previous float64
-
-		for c := range reporter {
-			if c > previous {
-				// todo: how to calculate how many characters can the line fit?
-				pb := utils.RenderProgressBar(suffix, c, pbScaling)
-				spinner.Update(pb)
-				previous = c
-			}
-		}
-	}()
-
-	if err := up.Sy.WaitForCompletion(ctx, up.Dev, reporter); err != nil {
-		analytics.TrackSyncError()
-		switch err {
-		case errors.ErrLostSyncthing:
-			return err
-		case errors.ErrInsufficientSpace:
-			return up.getInsufficientSpaceError(err)
-		default:
-			return errors.UserError{
-				E: err,
-				Hint: `Help us improve okteto by filing an issue in https://github.com/okteto/okteto/issues/new.
-    Please include the file generated by 'okteto doctor' if possible.
-    Then, try to run 'okteto down -v' + 'okteto up' again`,
-			}
-		}
-	}
-
-	// render to 100
-	spinner.Update(utils.RenderProgressBar(suffix, 100, pbScaling))
-
-	up.Sy.Type = "sendreceive"
-	up.Sy.IgnoreDelete = false
-	if err := up.Sy.UpdateConfig(); err != nil {
-		return err
-	}
-
-	go up.Sy.Monitor(ctx, up.Disconnect)
-	go up.Sy.MonitorStatus(ctx, up.Disconnect)
-	log.Infof("restarting syncthing to update sync mode to sendreceive")
-	return up.Sy.Restart(ctx)
-}
-
-func (up *upContext) cleanCommand(ctx context.Context) {
-	in := strings.NewReader("\n")
-	var out bytes.Buffer
-
-	cmd := "cat /var/okteto/bin/version.txt; cat /proc/sys/fs/inotify/max_user_watches; /var/okteto/bin/clean >/dev/null 2>&1"
-
-	err := exec.Exec(
-		ctx,
-		up.Client,
-		up.RestConfig,
-		up.Dev.Namespace,
-		up.Pod.Name,
-		up.Dev.Container,
-		false,
-		in,
-		&out,
-		os.Stderr,
-		[]string{"sh", "-c", cmd},
-	)
-
-	if err != nil {
-		log.Infof("failed to clean session: %s", err)
-	}
-
-	up.cleaned <- out.String()
-}
-
-func (up *upContext) runCommand(ctx context.Context) error {
-	log.Infof("starting remote command")
-	if err := config.UpdateStateFile(up.Dev, config.Ready); err != nil {
-		return err
-	}
-
-	if up.Dev.RemoteModeEnabled() {
-		return ssh.Exec(ctx, up.Dev.Interface, up.Dev.RemotePort, true, os.Stdin, os.Stdout, os.Stderr, up.Dev.Command.Values)
-	}
-
-	return exec.Exec(
-		ctx,
-		up.Client,
-		up.RestConfig,
-		up.Dev.Namespace,
-		up.Pod.Name,
-		up.Dev.Container,
-		true,
-		os.Stdin,
-		os.Stdout,
-		os.Stderr,
-		up.Dev.Command.Values,
-	)
-}
-
-func (up *upContext) checkOktetoStartError(ctx context.Context, msg string) error {
-	userID := pods.GetDevPodUserID(ctx, up.Dev, up.Client)
-	if up.Dev.PersistentVolumeEnabled() {
-		if userID != -1 && userID != *up.Dev.SecurityContext.RunAsUser {
-			return errors.UserError{
-				E: fmt.Errorf("User %d doesn't have write permissions for the %s directory", userID, up.Dev.MountPath),
-				Hint: fmt.Sprintf(`Set 'securityContext.runAsUser: %d' in your okteto manifest.
-    After that, run 'okteto down -v' to reset your development container and run 'okteto up' again`, userID),
-			}
-		}
-	} else {
-		if pods.OktetoDevPodMustBeRecreated(ctx, up.Dev, up.Client) {
-			return errors.ErrLostSyncthing
-		}
-	}
-
-	if len(up.Dev.Secrets) > 0 {
-		return errors.UserError{
-			E: fmt.Errorf(msg),
-			Hint: fmt.Sprintf(`Check your development container logs for errors: 'kubectl logs %s',
-    Check that your container can write to the destination path of your secrets.
-    Run 'okteto down -v' to reset your development container and try again`, up.Pod),
-		}
-	}
-	return errors.UserError{
-		E: fmt.Errorf(msg),
-		Hint: fmt.Sprintf(`Check your development container logs for errors: 'kubectl logs %s'.
-    Run 'okteto down -v' to reset your development container and try again`, up.Pod),
-	}
 }
 
 func (up *upContext) getInteractive() bool {
